test-requirements.txt
setuptools==56.0.0
black==22.3.0 ; python_version > '3.5'
coverage==4.5.4
flake8
git+https://github.com/ansible-network/pytest-ansible-network-integration.git
git+https://github.com/ansible-community/pytest-ansible-units.git
mock; python_version < '3'
pytest-xdist
yamllint
antsibull-changelog
